requirements.txt
biopython==1.76
matplotlib==2.2.5
networkx==2.2
numpy==1.16.6
pandas==0.23.0
#pygraphviz==1.5
scikit-learn==0.20.4
scipy==1.2.3
seaborn==0.9.1
statsmodels==0.10.0
tk==0.1.0

